+++ src/index.ts
@@ -1,2 +1,3 @@
 export * from "./errorboundary.js";
 export * from "./globalcrash.js";
+export * from "./i18n.js";